Enable the Cooja platform to build without using Cooja.
By moving the JNI functionality from platform.c to its own file, the
Cooja platform can be built similarly to other platforms, enabling
test compilation for Cooja without using Cooja.
